fix some lint issues

+++ protocol/governance/contracts/interfaces/IElectionInspectorModule.sol
@@ -9,35 +9,37 @@ interface IElectionInspectorModule {
     // ---------------------------------------
 
     /// @notice Returns the date in which the given epoch started
-    function getEpochStartDateForIndex(uint epochIndex) external view returns (uint64);
+    function getEpochStartDateForIndex(uint256 epochIndex) external view returns (uint64);
 
     /// @notice Returns the date in which the given epoch ended
-    function getEpochEndDateForIndex(uint epochIndex) external view returns (uint64);
+    function getEpochEndDateForIndex(uint256 epochIndex) external view returns (uint64);
 
     /// @notice Returns the date in which the Nomination period in the given epoch started
-    function getNominationPeriodStartDateForIndex(uint epochIndex) external view returns (uint64);
+    function getNominationPeriodStartDateForIndex(
+        uint256 epochIndex
+    ) external view returns (uint64);
 
     /// @notice Returns the date in which the Voting period in the given epoch started
-    function getVotingPeriodStartDateForIndex(uint epochIndex) external view returns (uint64);
+    function getVotingPeriodStartDateForIndex(uint256 epochIndex) external view returns (uint64);
 
     /// @notice Shows if a candidate was nominated in the given epoch
-    function wasNominated(address candidate, uint epochIndex) external view returns (bool);
+    function wasNominated(address candidate, uint256 epochIndex) external view returns (bool);
 
     /// @notice Returns a list of all nominated candidates in the given epoch
-    function getNomineesAtEpoch(uint epochIndex) external view returns (address[] memory);
+    function getNomineesAtEpoch(uint256 epochIndex) external view returns (address[] memory);
 
     /// @notice Returns if user has voted in the given election
     function hasVotedInEpoch(
         address user,
-        uint chainId,
-        uint epochIndex
+        uint256 chainId,
+        uint256 epochIndex
     ) external view returns (bool);
 
     function getCandidateVotesInEpoch(
         address candidate,
-        uint epochIndex
-    ) external view returns (uint);
+        uint256 epochIndex
+    ) external view returns (uint256);
 
     /// @notice Returns the winners of the given election
-    function getElectionWinnersInEpoch(uint epochIndex) external view returns (address[] memory);
+    function getElectionWinnersInEpoch(uint256 epochIndex) external view returns (address[] memory);
 }
