+++ protocol/governance/contracts/interfaces/external/ISnapshotRecord.sol
@@ -4,9 +4,9 @@ pragma solidity ^0.8.0;
 interface ISnapshotRecord {
     function currentPeriodId() external view returns (uint128);
 
-    function balanceOfOnPeriod(address account, uint periodId) external view returns (uint);
+    function balanceOfOnPeriod(address account, uint256 periodId) external view returns (uint256);
 
-    function totalSupplyOnPeriod(uint periodId) external view returns (uint);
+    function totalSupplyOnPeriod(uint256 periodId) external view returns (uint256);
 
     function takeSnapshot(uint128 id) external;
 }
