+++ protocol/governance/contracts/mocks/SnapshotRecordMock.sol
@@ -5,23 +5,26 @@ import {ISnapshotRecord} from "../interfaces/external/ISnapshotRecord.sol";
 
 contract SnapshotRecordMock is ISnapshotRecord {
     struct Period {
-        mapping(address => uint) balances;
+        mapping(address => uint256) balances;
     }
 
     mapping(uint128 => Period) private _periods;
-    mapping(uint128 => uint) private _totalSupplyOnPeriod;
+    mapping(uint128 => uint256) private _totalSupplyOnPeriod;
 
     // start at 1 to help the tests
     uint128 public currentPeriodId = 1;
 
-    function setBalanceOfOnPeriod(address user, uint balance, uint periodId) external {
+    function setBalanceOfOnPeriod(address user, uint256 balance, uint256 periodId) external {
         // solhint-disable-next-line numcast/safe-cast
         Period storage period = _periods[uint128(periodId)];
 
         period.balances[user] = balance;
     }
 
-    function balanceOfOnPeriod(address user, uint periodId) external view override returns (uint) {
+    function balanceOfOnPeriod(
+        address user,
+        uint256 periodId
+    ) external view override returns (uint256) {
         Period storage period = _periods[uint128(periodId)];
 
         return period.balances[user];
@@ -32,11 +35,11 @@ contract SnapshotRecordMock is ISnapshotRecord {
         // do nothing
     }
 
-    function setTotalSupplyOnPeriod(uint128 snapshotId, uint totalSupply) external {
+    function setTotalSupplyOnPeriod(uint128 snapshotId, uint256 totalSupply) external {
         _totalSupplyOnPeriod[snapshotId] = totalSupply;
     }
 
-    function totalSupplyOnPeriod(uint) external pure override returns (uint) {
+    function totalSupplyOnPeriod(uint256) external pure override returns (uint256) {
         return 0;
     }
 }
