+++ protocol/governance/contracts/modules/core/ElectionModule.sol
@@ -169,7 +169,7 @@ contract ElectionModule is IElectionModule, ElectionModuleSatellite, ElectionTal
         WormholeCrossChain.Data storage wh = WormholeCrossChain.load();
 
         uint64[] memory chains = wh.getSupportedNetworks();
-        for (uint i = 0; i < chains.length; i++) {
+        for (uint256 i = 0; i < chains.length; i++) {
             if (chains[i] == uint64(wh.getChainIdAt(0))) {
                 currentEpoch.nominationPeriodStartDate = newNominationPeriodStartDate;
                 currentEpoch.votingPeriodStartDate = newVotingPeriodStartDate;
@@ -389,7 +389,7 @@ contract ElectionModule is IElectionModule, ElectionModuleSatellite, ElectionTal
             WormholeCrossChain.Data storage wh = WormholeCrossChain.load();
 
             uint64[] memory chains = wh.getSupportedNetworks();
-            for (uint i = 0; i < chains.length; i++) {
+            for (uint256 i = 0; i < chains.length; i++) {
                 transmit(
                     wh,
                     uint16(chains[i]),
@@ -458,7 +458,7 @@ contract ElectionModule is IElectionModule, ElectionModuleSatellite, ElectionTal
         WormholeCrossChain.Data storage wh = WormholeCrossChain.load();
 
         uint64[] memory chains = wh.getSupportedNetworks();
-        for (uint i = 0; i < chains.length; i++) {
+        for (uint256 i = 0; i < chains.length; i++) {
             transmit(
                 wh,
                 uint16(chains[i]),
